Serious work to shrink the out-of-control ebayServer class, and break it up into functional helper classes.
Moved the ebayAuction parsing cleanup code into a cleaner of its own.

Moved the searching code into a search-specific class.

Moved the login logic into its own class, created an interface to talk to it, and use that everywhere instead of building it into ebayServer.

Extracted several string-processing blocks into the StringTools class, and disconnected them from the logic of the eBay server.

git-svn-id: svn://svn.jbidwatcher.com/jbidwatcher/trunk@58 b1acfa68-eb39-11db-b167-a3a8cd6b847e