Even more sweeping changes to fall back to a non-ebay.com eBay server (ebay.co.uk right now) for items which ebay.com is refusing to show.
It appears to work, and logging in even appears to work(!), but it needs real world testing, and real world bidding, buying, and sniping.

This is probably the largest set of deep modifications I've had to make in many years to support an eBay change.  It may make certain features harder in the future, but it's necessary to make it work again right now.

[#508]


git-svn-id: svn://svn.jbidwatcher.com/jbidwatcher/trunk@689 b1acfa68-eb39-11db-b167-a3a8cd6b847e