Clean up queueing so that the majority of queued events are strings.  This allows putting them in a persistent queue store.  The default 'enqueue' is string-only now.  Classes that want object-enqueueing can cast to PlainMessageQueue and call enqueueObject.  Syntactic vinegar.
Make many relatively simple value objects that get passed around in queues implement the bean specification, so they can be passed around as XML instead.  (This is done with 'enqueueBean', a terrible name, but dequeued objects are automatically unmarshalled back into the correct objects.)

Everything that passed around an AuctionEntry object to a queue now just passes the identifier, and the queue processor gets the entry from EntryCorral.  There may be a better way to do this, but for now, this works.

The TimeQueueManager does unpleasant things to manage the differences between different kinds of payloads that need to be delivered.  Possibly wrong things, also.

Canceling and setting snipes go directly to the server, instead of being queued, since it's not a long-running operation.  (Lowering the number of moving parts in sniping operations.)

Thumbnail lookup and handling just involves a String URL, which is turned into real URL in the manager.  This allows the thumbnail manager to be the only one that actually downloads the images, instead of the ebayAuction object.

The code to clear out multiple TIMECHECK operations is removed.

There is a custom Searcher serialize/deserialize method that calls on SearchManager.getSearchById() to only load the right item.  Theoretically I could do this for enqueuing AuctionEntry objects (on deserialize, pull them from EntryCorral) also, but...not right now. 

I hate making huge checkins like this, but once I started making deep changes to the queueing infrastructure, I couldn't check it in until it was all at least stable and working.


git-svn-id: svn://svn.jbidwatcher.com/jbidwatcher/trunk@893 b1acfa68-eb39-11db-b167-a3a8cd6b847e