Remove the backup server concept in favor of user-chosen sole server.
If the user hasn't chosen, then we go with the US server (ebay.com) if their default locale is US, CA, or an empty locale string.  Otherwise we go with ebay.co.uk.  You can configure the alternate server with `ebay.alternate` in the advanced configuration, but there has to be a properties file for it.